Updated error handling for state mutator. Updated godoc. Renamed files
